Separated the installed_packages function
this allows for all providers to test if something is installed.
