+++ metadata.rb
@@ -4,6 +4,6 @@ maintainer_email 'cookbooks@getchef.com'
 license          'Apache 2.0'
 description      'Provides a set of useful Windows-specific primitives.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.36.1'
+version          '1.36.2'
 supports         'windows'
 depends          'chef_handler'
